Extract sheet close button
